+++ lib/src/state_machine/mp_th2_file_edit_state_machine/mixins/mp_th2_file_edit_page_simplify_line_mixin.dart
@@ -1,7 +1,4 @@
-import 'package:flutter/services.dart';
-import 'package:mapiah/src/auxiliary/mp_interaction_aux.dart';
-import 'package:mapiah/src/controllers/th2_file_edit_element_edit_controller.dart';
-import 'package:mapiah/src/state_machine/mp_th2_file_edit_state_machine/mp_th2_file_edit_state.dart';
+part of '../mp_th2_file_edit_state.dart';
 
 mixin MPTH2FileEditPageSimplifyLineMixin on MPTH2FileEditState {
   bool onKeyLDownEvent(KeyDownEvent event) {