+++ lib/src/state_machine/mp_th2_file_edit_state_machine/mp_th2_file_edit_state_add_line_to_area.dart
@@ -2,6 +2,7 @@ part of 'mp_th2_file_edit_state.dart';
 
 class MPTH2FileEditStateAddLineToArea extends MPTH2FileEditState
     with
+        MPTH2FileEditPageAltClickMixin,
         MPTH2FileEditPageStateAddLineToAreaMixin,
         MPTH2FileEditStateClearSelectionOnExitMixin,
         MPTH2FileEditStateMoveCanvasMixin {
@@ -41,6 +42,10 @@ class MPTH2FileEditStateAddLineToArea extends MPTH2FileEditState
 
   @override
   Future<void> onPrimaryButtonClick(PointerUpEvent event) async {
+    if (onAltPrimaryButtonClick(event)) {
+      return Future.value();
+    }
+
     final MPCommand? addLineToAreaCommand = await getAddLineToAreaCommand(
       event: event,
       th2FileEditController: th2FileEditController,