+++ CHANGELOG.md
@@ -6,6 +6,7 @@
   * English install instructions.
   * Right click overlay windows dragging support.
   * Portuguese help pages.
+  * _Alt+click_ to select non active scrap.
 * Fixed bugs:
   * When hidden one of the background images in cave.th2 (demo.zip) the other image moves.
   * When dragging (right click) sketch, unselected sketchs are not dragged.